trying to get exported symbols right

+++ src/types/jack/ringbuffer.c
@@ -42,7 +42,7 @@ jack_ringbuffer_create (size_t sz)
 		return NULL;
 	}
 	
-	for (power_of_two = 1; 1 << power_of_two < sz; power_of_two++);
+	for (power_of_two = (size_t)1; (size_t)1 << power_of_two < sz; power_of_two++);
 	
 	rb->size = 1 << power_of_two;
 	rb->size_mask = rb->size;
